Time challengde, own prefab use, scene story, prefab instantiate
Added a time challenge so after dialouge with NPC we instantiate 3 mystic items in the scene, we have 3 seconds to pick 3 up. There is a "SHORT" scene story where the npc says we need to pick them up to save villagers in another world. We win or lose depending on what.

Also we use the SpecialItemPrefab and also I use the sliders and a button to reuse.